refactor(shape-plugin): extract TSX logic into custom hooks
