+++ plugins/shape-plugin/src/ui/components/build-config/SourceInvalidGeometryFilterCard.tsx
@@ -17,9 +17,10 @@ import {
 import { BuildConfigSectionTitle, getBuildConfigHoverCardSx } from '@hierarchidb/ui-accordion-config';
 import { useTranslation } from '~/ui/i18n';
 import type { ShapeBuildConfig } from '~/common/types/index';
-import { applyBuildConfigPatch } from '~/common/types/index';
-import type { ChangeEvent } from 'react';
-import { useCallback, useMemo } from 'react';
+import {
+  useSourceGeometryIntakeGuardCardView,
+  useTileEmitInvalidGeometryFilterCardView,
+} from './useSourceInvalidGeometryFilterCardView.js';
 
 type Props = {
   config: ShapeBuildConfig;
@@ -27,41 +28,6 @@ type Props = {
   disabled?: boolean;
   disableHoverLift?: boolean;
 };
-
-type InvalidGeometryFilterState = {
-  area: boolean;
-  lineLength: boolean;
-  maxEdgeLength: boolean;
-  selfIntersection: boolean;
-  triangleRingRatio: boolean;
-};
-
-type GeometryIntakeGuardState = {
-  validationLevel: 'off' | 'basic' | 'strict';
-  dedupeEpsilon: number;
-  minRingAreaThreshold: number;
-  normalizeRingOrientation: boolean;
-  keepBaselineSnapshot: boolean;
-};
-
-const resolveGeometryIntakeGuard = (config: ShapeBuildConfig): GeometryIntakeGuardState => {
-  const guard = config.sourceConfig.geometryIntakeGuard;
-  return {
-    validationLevel: guard?.validationLevel ?? 'off',
-    dedupeEpsilon: guard?.dedupeEpsilon ?? 0.000001,
-    minRingAreaThreshold: guard?.minRingAreaThreshold ?? 0,
-    normalizeRingOrientation: guard?.normalizeRingOrientation ?? true,
-    keepBaselineSnapshot: guard?.keepBaselineSnapshot ?? true,
-  };
-};
-
-const resolveInvalidGeometryFilter = (config: ShapeBuildConfig): InvalidGeometryFilterState => ({
-  area: config.tileEmitConfig.invalidGeometryFilter?.area ?? false,
-  lineLength: config.tileEmitConfig.invalidGeometryFilter?.lineLength ?? false,
-  maxEdgeLength: config.tileEmitConfig.invalidGeometryFilter?.maxEdgeLength ?? false,
-  selfIntersection: config.tileEmitConfig.invalidGeometryFilter?.selfIntersection ?? false,
-  triangleRingRatio: config.tileEmitConfig.invalidGeometryFilter?.triangleRingRatio ?? false,
-});
 
 export const SourceGeometryIntakeGuardCard: React.FC<Props> = ({
   config,
@@ -71,22 +37,14 @@ export const SourceGeometryIntakeGuardCard: React.FC<Props> = ({
   const { t } = useTranslation();
   const guardCardDisabled = true;
   const hoverCardSx = getBuildConfigHoverCardSx(guardCardDisabled, disableHoverLift);
-  const resolvedGuard = useMemo(() => resolveGeometryIntakeGuard(config), [config]);
-
-  const updateGuard = useCallback((partial: Partial<GeometryIntakeGuardState>): void => {
-    onChange((prevConfig) => {
-      const current = resolveGeometryIntakeGuard(prevConfig);
-      return applyBuildConfigPatch(prevConfig, {
-        sourceConfig: {
-          ...prevConfig.sourceConfig,
-          geometryIntakeGuard: {
-            ...current,
-            ...partial,
-          },
-        },
-      });
-    });
-  }, [onChange]);
+  const {
+    resolvedGuard,
+    handleValidationLevelChange,
+    handleDedupeEpsilonChange,
+    handleMinRingAreaThresholdChange,
+    handleNormalizeRingOrientationChange,
+    handleKeepBaselineSnapshotChange,
+  } = useSourceGeometryIntakeGuardCardView(config, onChange);
 
   const isDisabled = true;
 
@@ -122,11 +80,7 @@ export const SourceGeometryIntakeGuardCard: React.FC<Props> = ({
                 exclusive
                 disabled
                 value={resolvedGuard.validationLevel}
-                onChange={(_event, value) => {
-                  if (value === null) return;
-                  if (value !== 'off' && value !== 'basic' && value !== 'strict') return;
-                  updateGuard({ validationLevel: value });
-                }}
+                onChange={handleValidationLevelChange}
               >
                 <ToggleButton value="off">{t('processing.source.geometryIntakeGuard.level.off', 'off')}</ToggleButton>
                 <ToggleButton value="basic">{t('processing.source.geometryIntakeGuard.level.basic', 'basic')}</ToggleButton>
@@ -141,11 +95,7 @@ export const SourceGeometryIntakeGuardCard: React.FC<Props> = ({
               label={t('processing.source.geometryIntakeGuard.dedupeEpsilon', 'Duplicate vertex epsilon')}
               value={resolvedGuard.dedupeEpsilon}
               disabled
-              onChange={(event) => {
-                const value = Number(event.target.value);
-                if (!Number.isFinite(value)) return;
-                updateGuard({ dedupeEpsilon: Math.max(0, value) });
-              }}
+              onChange={handleDedupeEpsilonChange}
             />
           </Grid>
           <Grid size={{ xs: 12, md: 4 }}>
@@ -155,11 +105,7 @@ export const SourceGeometryIntakeGuardCard: React.FC<Props> = ({
               label={t('processing.source.geometryIntakeGuard.minRingAreaThreshold', 'Minimum ring area threshold')}
               value={resolvedGuard.minRingAreaThreshold}
               disabled
-              onChange={(event) => {
-                const value = Number(event.target.value);
-                if (!Number.isFinite(value)) return;
-                updateGuard({ minRingAreaThreshold: Math.max(0, value) });
-              }}
+              onChange={handleMinRingAreaThresholdChange}
             />
           </Grid>
         </Grid>
@@ -167,14 +113,14 @@ export const SourceGeometryIntakeGuardCard: React.FC<Props> = ({
         <Grid container spacing={1.5}>
           <Grid size={{ xs: 12, md: 6 }}>
             <FormControlLabel
-              control={<Switch checked={resolvedGuard.normalizeRingOrientation} onChange={(event) => updateGuard({ normalizeRingOrientation: event.target.checked })} />}
+              control={<Switch checked={resolvedGuard.normalizeRingOrientation} onChange={handleNormalizeRingOrientationChange} />}
               disabled={isDisabled}
               label={t('processing.source.geometryIntakeGuard.normalizeRingOrientation', 'Normalize ring orientation')}
             />
           </Grid>
           <Grid size={{ xs: 12, md: 6 }}>
             <FormControlLabel
-              control={<Switch checked={resolvedGuard.keepBaselineSnapshot} onChange={(event) => updateGuard({ keepBaselineSnapshot: event.target.checked })} />}
+              control={<Switch checked={resolvedGuard.keepBaselineSnapshot} onChange={handleKeepBaselineSnapshotChange} />}
               disabled={isDisabled}
               label={t('processing.source.geometryIntakeGuard.keepBaselineSnapshot', 'Keep baseline snapshot for anomaly scoring')}
             />
@@ -193,67 +139,18 @@ export const TileEmitInvalidGeometryFilterCard: React.FC<Props> = ({
 }) => {
   const { t } = useTranslation();
   const hoverCardSx = getBuildConfigHoverCardSx(disabled, disableHoverLift);
-  const resolved = useMemo(() => resolveInvalidGeometryFilter(config), [config]);
-
-  const updateFilter = useCallback((partial: Partial<InvalidGeometryFilterState>): void => {
-    onChange((prevConfig) => {
-      const current = resolveInvalidGeometryFilter(prevConfig);
-      return applyBuildConfigPatch(prevConfig, {
-        tileEmitConfig: {
-          ...prevConfig.tileEmitConfig,
-          invalidGeometryFilter: {
-            ...current,
-            ...partial,
-          },
-        },
-      });
-    });
-  }, [onChange]);
-
-  type SwitchItem = {
-    checked: boolean;
-    disabled: boolean;
-    label: string;
-    onChange: (event: ChangeEvent<HTMLInputElement>) => void;
-  };
-
-  const isDisabled = Boolean(disabled);
-  const switchGroups: Array<Array<SwitchItem>> = [
-    [
-      {
-        checked: resolved.selfIntersection,
-        disabled: isDisabled,
-        label: t('processing.tileEmit.invalidGeometryFilter.selfIntersection', 'Self intersection'),
-        onChange: (event) => updateFilter({ selfIntersection: event.target.checked }),
-      },
-      {
-        checked: resolved.triangleRingRatio,
-        disabled: isDisabled,
-        label: t('processing.tileEmit.invalidGeometryFilter.triangleRingRatio', 'Triangle ring ratio'),
-        onChange: (event) => updateFilter({ triangleRingRatio: event.target.checked }),
-      },
-    ],
-    [
-      {
-        checked: resolved.area,
-        disabled: isDisabled,
-        label: t('processing.tileEmit.invalidGeometryFilter.area', 'Area'),
-        onChange: (event) => updateFilter({ area: event.target.checked }),
-      },
-      {
-        checked: resolved.lineLength,
-        disabled: isDisabled,
-        label: t('processing.tileEmit.invalidGeometryFilter.lineLength', 'Line length'),
-        onChange: (event) => updateFilter({ lineLength: event.target.checked }),
-      },
-      {
-        checked: resolved.maxEdgeLength,
-        disabled: isDisabled,
-        label: t('processing.tileEmit.invalidGeometryFilter.maxEdgeLength', 'Max edge length'),
-        onChange: (event) => updateFilter({ maxEdgeLength: event.target.checked }),
-      },
-    ],
-  ];
+  const { switchGroups } = useTileEmitInvalidGeometryFilterCardView(
+    config,
+    onChange,
+    disabled,
+    {
+      selfIntersection: t('processing.tileEmit.invalidGeometryFilter.selfIntersection', 'Self intersection'),
+      triangleRingRatio: t('processing.tileEmit.invalidGeometryFilter.triangleRingRatio', 'Triangle ring ratio'),
+      area: t('processing.tileEmit.invalidGeometryFilter.area', 'Area'),
+      lineLength: t('processing.tileEmit.invalidGeometryFilter.lineLength', 'Line length'),
+      maxEdgeLength: t('processing.tileEmit.invalidGeometryFilter.maxEdgeLength', 'Max edge length'),
+    },
+  );
 
   return (
     <Paper variant="outlined" sx={{ p: 2, ...hoverCardSx }}>
