+++ plugins/shape-plugin/src/ui/components/build-progress/ShapeBuildProgressPanel/useShapeBuildProgressPanelController/useShapeBuildProgressPanelControllerOverlay/useShapeBuildProgressPanelControllerOverlaySections.ts
@@ -4,7 +4,7 @@ import type { ShapeBuildProgressPanelControllerBaseResult } from '~/ui/component
 import {
   ShapeBuildProgressPanelOverlaySectionContent,
   ShapeBuildProgressPanelOverlaySectionProgressContent,
-} from './useShapeBuildProgressPanelControllerOverlaySectionsView.js';
+} from './ShapeBuildProgressPanelControllerOverlaySectionsView.js';
 import { resolveStageAliasArray, resolveStageAliasValue } from '~/ui/components/build-progress/stageIdAliases';
 
 type StageRecord = Record<string, ReactNode>;
