+++ plugins/shape-plugin/src/ui/components/build-progress/tile-emit/TransposedFeaturePreviewCard.tsx
@@ -1,7 +1,6 @@
 import type { Feature, Geometry, Position } from 'geojson';
 import { Box, Paper, Stack } from '@mui/material';
-import { useCallback, useMemo, useRef, useState } from 'react';
-import type { PointerEvent, WheelEvent } from 'react';
+import { useTransposedFeaturePreviewCardView } from './useTransposedFeaturePreviewCardView.js';
 
 export type TileBBox = {
   minX: number;
@@ -37,18 +36,6 @@ const lonLatToMercator = (lon: number, lat: number): Point | null => {
   const x = lon * 20037508.34 / 180;
   const y = Math.log(Math.tan((90 + clampedLat) * rad / 2)) * 20037508.34 / Math.PI;
   return { x, y };
-};
-
-const toMercatorBBox = (bbox: TileBBox): MercatorBBox | null => {
-  const min = lonLatToMercator(bbox.minX, bbox.minY);
-  const max = lonLatToMercator(bbox.maxX, bbox.maxY);
-  if (!min || !max) return null;
-  return {
-    minX: Math.min(min.x, max.x),
-    minY: Math.min(min.y, max.y),
-    maxX: Math.max(min.x, max.x),
-    maxY: Math.max(min.y, max.y),
-  };
 };
 
 const projectPoint = (coord: Position, view: MercatorBBox, width: number, height: number): Point | null => {
@@ -113,92 +100,25 @@ export const TransposedFeaturePreviewCard = ({
   selectedFeatureIds,
   onMouseLeave,
 }: TransposedFeaturePreviewCardProps) => {
-  const containerRef = useRef<HTMLDivElement | null>(null);
-  const [scale, setScale] = useState(1);
-  const [offset, setOffset] = useState<Point>({ x: 0, y: 0 });
-  const [dragging, setDragging] = useState(false);
-  const [lastPoint, setLastPoint] = useState<Point | null>(null);
-
-  const view = useMemo(() => toMercatorBBox(bufferBBox), [bufferBBox]);
-  const width = 1000;
-  const height = 1000;
-  const tileView = useMemo(() => toMercatorBBox(tileBBox), [tileBBox]);
-  const tileTopLeft = view && tileView
-    ? projectPoint([tileBBox.minX, tileBBox.maxY], view, width, height)
-    : null;
-  const tileBottomRight = view && tileView
-    ? projectPoint([tileBBox.maxX, tileBBox.minY], view, width, height)
-    : null;
-  const tileRect = tileTopLeft && tileBottomRight
-    ? {
-      x: tileTopLeft.x,
-      y: tileTopLeft.y,
-      width: tileBottomRight.x - tileTopLeft.x,
-      height: tileBottomRight.y - tileTopLeft.y,
-    }
-    : null;
-
-  const handleWheel = useCallback((event: WheelEvent<HTMLDivElement>) => {
-    event.preventDefault();
-    const rect = containerRef.current?.getBoundingClientRect();
-    if (!rect) return;
-    if (rect.width <= 0 || rect.height <= 0) return;
-    const pointerX = event.clientX - rect.left;
-    const pointerY = event.clientY - rect.top;
-    const pointerSvgX = (pointerX / rect.width) * width;
-    const pointerSvgY = (pointerY / rect.height) * height;
-    const zoomFactor = event.deltaY < 0 ? 1.1 : 0.9;
-    setScale((prev) => {
-      const next = Math.min(8, Math.max(0.5, prev * zoomFactor));
-      if (next === prev) return prev;
-      const ratio = prev / next;
-      setOffset((prevOffset) => ({
-        x: (prevOffset.x + pointerSvgX) * ratio - pointerSvgX,
-        y: (prevOffset.y + pointerSvgY) * ratio - pointerSvgY,
-      }));
-      return next;
-    });
-  }, [height, width]);
-
-  const handlePointerDown = useCallback((event: PointerEvent<HTMLDivElement>) => {
-    event.currentTarget.setPointerCapture(event.pointerId);
-    setDragging(true);
-    setLastPoint({ x: event.clientX, y: event.clientY });
-  }, []);
-
-  const handlePointerMove = useCallback((event: PointerEvent<HTMLDivElement>) => {
-    if (!dragging || !lastPoint) return;
-    const dx = event.clientX - lastPoint.x;
-    const dy = event.clientY - lastPoint.y;
-    setOffset((prev) => ({ x: prev.x + dx, y: prev.y + dy }));
-    setLastPoint({ x: event.clientX, y: event.clientY });
-  }, [dragging, lastPoint]);
-
-  const handlePointerUp = useCallback((event?: PointerEvent<HTMLDivElement>) => {
-    if (event) {
-      event.currentTarget.releasePointerCapture(event.pointerId);
-    }
-    setDragging(false);
-    setLastPoint(null);
-  }, []);
-
-  const resetView = useCallback(() => {
-    setScale(1);
-    setOffset({ x: 0, y: 0 });
-  }, []);
-
-  const handleMouseLeave = useCallback(() => {
-    setDragging(false);
-    setLastPoint(null);
-    resetView();
-    onMouseLeave?.();
-  }, [onMouseLeave, resetView]);
-
-  const handlePointerLeave = useCallback((event: PointerEvent<HTMLDivElement>) => {
-    handlePointerUp(event);
-    resetView();
-    onMouseLeave?.();
-  }, [handlePointerUp, onMouseLeave, resetView]);
+  const {
+    containerRef,
+    view,
+    width,
+    height,
+    tileRect,
+    scale,
+    offset,
+    handleWheel,
+    handlePointerDown,
+    handlePointerMove,
+    handlePointerUp,
+    handlePointerLeave,
+    handleMouseLeave,
+  } = useTransposedFeaturePreviewCardView({
+    tileBBox,
+    bufferBBox,
+    onMouseLeave,
+  });
 
   return (
     <Paper variant="outlined" sx={{ p: 1.25 }}>
